requirements.txt
Django==1.11.4
django-rest-auth[with_social]==0.9.1
djangorestframework==3.6.4
gunicorn==19.7.0
psycopg2==2.7.3.1
